requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
langchain-core
streamlit